Allow specifying ints as some pointers
glVertexAttribPointer and glDrawelements take a "pointer" parameter, which
typically is 0. Previously, glcaml segfaulted as it was expecting a proper
String or bigarray there: now allow specifying just a raw int.

Surely there are other functions where this change would be healthy. I will
probably fix more as the need arises.